moving debug command into own class
